+++ src/validation.rs
@@ -18,7 +18,7 @@ use crypto::Algorithm;
 /// let validation = Validation::default();
 ///
 /// // Changing one parameter
-/// let mut validation = Validation {leeway: 1000 * 60, ..Default::default()};
+/// let mut validation = Validation {leeway: 60, ..Default::default()};
 ///
 /// // Setting audience
 /// let mut validation = Validation::default();
@@ -27,17 +27,11 @@ use crypto::Algorithm;
 /// ```
 #[derive(Debug, Clone, PartialEq)]
 pub struct Validation {
-    /// Add some leeway (in ms) to the `exp`, `iat` and `nbf` validation to
+    /// Add some leeway (in seconds) to the `exp`, `iat` and `nbf` validation to
     /// account for clock skew.
     ///
     /// Defaults to `0`.
     pub leeway: i64,
-    /// Whether to actually validate the signature of the token.
-    ///
-    /// WARNING: only set that to false if you know what you are doing.
-    ///
-    /// Defaults to `true`.
-    pub validate_signature: bool,
     /// Whether to validate the `exp` field.
     ///
     /// It will return an error if the time in the `exp` field is past.
@@ -92,8 +86,6 @@ impl Default for Validation {
     fn default() -> Validation {
         Validation {
             leeway: 0,
-
-            validate_signature: true,
 
             validate_exp: true,
             validate_iat: true,